chore: change the architecture
