+++ frontend_app/App.vue
