+++ frontend_app/pages/AboutPage.Vue
