+++ frontend_app/pages/HomePage copy.vue
