+++ frontend_app/pages/HomePage.vue
