+++ frontend_app/static/global.css
