+++ package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "webpack serve --config config/webpack.dev.js",
+    "dev": "webpack serve --config webpack/webpack.dev.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "prod": "webpack --config config/webpack.prod.js"
+    "prod": "webpack --config webpack/webpack.prod.js"
   },
   "author": "",
   "license": "ISC",
