--- views/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <h1>Test</h1>
-    <div id="app"></div>
-    <script src="/assets/js/main.bundle.js"></script>
-  </body>
-</html>
